Implement save validation. Version check to come

+++ include/game_variables.h
@@ -34,7 +34,7 @@ typedef struct
 
     // BY DEFAULT IS SET TO 1, but if changed to 2 or more, should speed up all (or most) of the
     // game aspects that should be sped up by speed, as in the original game.
-    int game_speed;
+    u8 game_speed;
 
     bool high_contrast;
